extract recref_adapter.py
