+++ location/views.py
@@ -11,7 +11,7 @@ from django.views.generic import ListView
 from core.constant import REL_TYPE_COMMENT_REFERS_TO, REL_TYPE_IS_RELATED_TO
 from core.forms import CommentForm, ResourceForm
 from core.helper import view_utils, renderer_utils, query_utils, download_csv_utils
-from core.helper.common_recref_adapter import RecrefFormAdapter, TargetCommentRecrefAdapter, TargetResourceRecrefAdapter
+from core.helper.common_recref_adapter import RecrefFormAdapter
 from core.helper.model_utils import RecordTracker
 from core.helper.renderer_utils import CompactSearchResultsRenderer
 from core.helper.view_components import DownloadCsvHandler
@@ -20,6 +20,7 @@ from core.models import Recref
 from location import location_utils
 from location.forms import LocationForm, GeneralSearchFieldset
 from location.models import CofkUnionLocation, CofkLocationCommentMap, CofkLocationResourceMap
+from location.recref_adapter import LocationCommentRecrefAdapter, LocationResourceRecrefAdapter
 
 log = logging.getLogger(__name__)
 FormOrFormSet = Union[BaseForm, BaseFormSet]
@@ -302,22 +303,6 @@ class LocationCommentFormsetHandler(RecrefFormsetHandler):
         return CofkLocationCommentMap.objects.filter(location=parent, comment=target).first()
 
 
-class LocationCommentRecrefAdapter(TargetCommentRecrefAdapter):
-    def __init__(self, parent):
-        self.parent: CofkUnionLocation = parent
-
-    def recref_class(self) -> Type[Recref]:
-        return CofkLocationCommentMap
-
-    def set_parent_target_instance(self, recref, parent, target):
-        recref: CofkLocationCommentMap
-        recref.location = parent
-        recref.comment = target
-
-    def find_recref_records(self, rel_type):
-        return self.find_recref_records_by_related_manger(self.parent.cofklocationcommentmap_set, rel_type)
-
-
 class LocationResourceFormsetHandler(RecrefFormsetHandler):
     def create_recref_adapter(self, parent) -> RecrefFormAdapter:
         return LocationResourceRecrefAdapter(parent)
@@ -326,17 +311,3 @@ class LocationResourceFormsetHandler(RecrefFormsetHandler):
         return CofkLocationResourceMap.objects.filter(location=parent, resource=target).first()
 
 
-class LocationResourceRecrefAdapter(TargetResourceRecrefAdapter):
-    def __init__(self, parent):
-        self.parent: CofkUnionLocation = parent
-
-    def recref_class(self) -> Type[Recref]:
-        return CofkLocationResourceMap
-
-    def set_parent_target_instance(self, recref, parent, target):
-        recref: CofkLocationResourceMap
-        recref.location = parent
-        recref.resource = target
-
-    def find_recref_records(self, rel_type):
-        return self.find_recref_records_by_related_manger(self.parent.cofklocationresourcemap_set, rel_type)
