+++ person/views.py
@@ -1,5 +1,4 @@
 import logging
-from abc import ABC
 from typing import Callable, Iterable, Type, Optional, Any, NoReturn
 
 from django.contrib.auth.decorators import login_required
@@ -13,10 +12,8 @@ from django.shortcuts import render, redirect, get_object_or_404
 from core import constant
 from core.constant import REL_TYPE_COMMENT_REFERS_TO, REL_TYPE_IS_RELATED_TO
 from core.forms import CommentForm, ResourceForm, LocRecrefForm, PersonRecrefForm
-from core.helper import renderer_utils, view_utils, query_utils, download_csv_utils, recref_utils, form_utils, \
-    model_utils
-from core.helper.common_recref_adapter import RecrefFormAdapter, TargetCommentRecrefAdapter, \
-    TargetResourceRecrefAdapter, TargetPersonRecrefAdapter
+from core.helper import renderer_utils, view_utils, query_utils, download_csv_utils, recref_utils, form_utils
+from core.helper.common_recref_adapter import RecrefFormAdapter
 from core.helper.renderer_utils import CompactSearchResultsRenderer
 from core.helper.view_components import DownloadCsvHandler
 from core.helper.view_utils import CommonInitFormViewTemplate, ImageHandler, BasicSearchView, FullFormHandler, \
@@ -26,8 +23,9 @@ from location.models import CofkUnionLocation
 from person import person_utils
 from person.forms import PersonForm, GeneralSearchFieldset, PersonOtherRecrefForm
 from person.models import CofkUnionPerson, CofkPersonLocationMap, CofkPersonPersonMap, create_person_id, \
-    CofkPersonCommentMap, CofkPersonResourceMap, CofkPersonRoleMap
-from uploader.models import CofkUnionRoleCategory
+    CofkPersonCommentMap, CofkPersonResourceMap
+from person.recref_adapter import PersonCommentRecrefAdapter, PersonResourceRecrefAdapter, PersonRoleRecrefAdapter, \
+    ActivePersonRecrefAdapter, PassivePersonRecrefAdapter
 
 log = logging.getLogger(__name__)
 
@@ -271,7 +269,7 @@ class PersonSearchView(LoginRequiredMixin, BasicSearchView):
     @property
     def sort_by_choices(self) -> list[tuple[str, str]]:
         return [
-            #('names_and_titles', 'Names and titles / roles',), TODO this is in a view cofk_union_person_view
+            # ('names_and_titles', 'Names and titles / roles',), TODO this is in a view cofk_union_person_view
             ('date_of_birth', 'Born',),
             ('date_of_death', 'Died',),
             ('flourished', 'Flourished',),
@@ -418,103 +416,9 @@ class PersonCommentFormsetHandler(RecrefFormsetHandler):
         return CofkPersonCommentMap.objects.filter(person=parent, comment=target).first()
 
 
-class PersonCommentRecrefAdapter(TargetCommentRecrefAdapter):
-    def __init__(self, parent):
-        self.parent: CofkUnionPerson = parent
-
-    def recref_class(self) -> Type[Recref]:
-        return CofkPersonCommentMap
-
-    def set_parent_target_instance(self, recref, parent, target):
-        recref: CofkPersonCommentMap
-        recref.person = parent
-        recref.comment = target
-
-    def find_recref_records(self, rel_type):
-        return self.find_recref_records_by_related_manger(self.parent.cofkpersoncommentmap_set, rel_type)
-
-
 class PersonResourceFormsetHandler(RecrefFormsetHandler):
     def create_recref_adapter(self, parent) -> RecrefFormAdapter:
         return PersonResourceRecrefAdapter(parent)
 
     def find_org_recref_fn(self, parent, target) -> Recref | None:
         return CofkPersonResourceMap.objects.filter(person=parent, resource=target).first()
-
-
-class PersonResourceRecrefAdapter(TargetResourceRecrefAdapter):
-    def __init__(self, parent):
-        self.parent: CofkUnionPerson = parent
-
-    def recref_class(self) -> Type[Recref]:
-        return CofkPersonResourceMap
-
-    def set_parent_target_instance(self, recref, parent, target):
-        recref: CofkPersonResourceMap
-        recref.person = parent
-        recref.resource = target
-
-    def find_recref_records(self, rel_type):
-        return self.find_recref_records_by_related_manger(self.parent.cofkpersonresourcemap_set, rel_type)
-
-
-class PersonRoleRecrefAdapter(RecrefFormAdapter):
-    def __init__(self, parent):
-        self.parent: CofkUnionPerson = parent
-
-    def find_target_display_name_by_id(self, target_id):
-        target = self.find_target_instance(target_id)
-        return target and target.role_category_desc
-
-    def recref_class(self) -> Type[Recref]:
-        return CofkPersonRoleMap
-
-    def find_target_instance(self, target_id):
-        return model_utils.get_safe(CofkUnionRoleCategory, pk=target_id)
-
-    def set_parent_target_instance(self, recref, parent, target):
-        recref: CofkPersonRoleMap
-        recref.person = parent
-        recref.role = target
-
-    def find_recref_records(self, rel_type):
-        return self.find_recref_records_by_related_manger(self.parent.cofkpersonrolemap_set, rel_type)
-
-    def target_id_name(self):
-        return 'role_id'
-
-
-class PersonPersonRecrefAdapter(TargetPersonRecrefAdapter, ABC):
-    def __init__(self, parent=None):
-        self.parent: CofkUnionPerson = parent
-
-    def recref_class(self) -> Type[Recref]:
-        return CofkPersonPersonMap
-
-
-class ActivePersonRecrefAdapter(PersonPersonRecrefAdapter):
-
-    def set_parent_target_instance(self, recref, parent, target):
-        recref: CofkPersonPersonMap
-        recref.person = parent
-        recref.related = target
-
-    def find_recref_records(self, rel_type):
-        return self.find_recref_records_by_related_manger(self.parent.active_relationships, rel_type)
-
-    def target_id_name(self):
-        return 'related_id'
-
-
-class PassivePersonRecrefAdapter(PersonPersonRecrefAdapter):
-
-    def set_parent_target_instance(self, recref, parent, target):
-        recref: CofkPersonPersonMap
-        recref.person = target
-        recref.related = parent
-
-    def find_recref_records(self, rel_type):
-        return self.find_recref_records_by_related_manger(self.parent.passive_relationships, rel_type)
-
-    def target_id_name(self):
-        return 'person_id'
